+++ test/OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests/OtlpLogExporterTests.cs
@@ -17,6 +17,7 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.Linq;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
@@ -27,12 +28,15 @@ using OpenTelemetry.Proto.Collector.Logs.V1;
 using OpenTelemetry.Tests;
 using OpenTelemetry.Trace;
 using Xunit;
+using OtlpCommon = OpenTelemetry.Proto.Common.V1;
 using OtlpLogs = OpenTelemetry.Proto.Logs.V1;
 
 namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
 {
     public class OtlpLogExporterTests : Http2UnencryptedSupportTests
     {
+        private static readonly SdkLimitOptions DefaultSdkLimitOptions = new();
+
         [Fact]
         public void AddOtlpLogExporterReceivesAttributesWithParseStateValueSetToFalse()
         {
@@ -187,7 +191,7 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
             Assert.Single(logRecords);
 
             var logRecord = logRecords[0];
-            var otlpLogRecord = logRecord.ToOtlpLog();
+            var otlpLogRecord = logRecord.ToOtlpLog(DefaultSdkLimitOptions);
 
             Assert.NotNull(otlpLogRecord);
             Assert.Equal("Hello from tomato 2.99.", otlpLogRecord.Body.StringValue);
@@ -224,7 +228,7 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
             Assert.Single(logRecords);
 
             var logRecord = logRecords[0];
-            var otlpLogRecord = logRecord.ToOtlpLog();
+            var otlpLogRecord = logRecord.ToOtlpLog(DefaultSdkLimitOptions);
             Assert.NotNull(otlpLogRecord);
             Assert.Single(otlpLogRecord.Attributes);
 
@@ -238,7 +242,7 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
             Assert.Single(logRecords);
 
             logRecord = logRecords[0];
-            otlpLogRecord = logRecord.ToOtlpLog();
+            otlpLogRecord = logRecord.ToOtlpLog(DefaultSdkLimitOptions);
             Assert.NotNull(otlpLogRecord);
             Assert.Empty(otlpLogRecord.Attributes);
         }
@@ -263,7 +267,7 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
             Assert.Single(logRecords);
 
             var logRecord = logRecords[0];
-            var otlpLogRecord = logRecord.ToOtlpLog();
+            var otlpLogRecord = logRecord.ToOtlpLog(DefaultSdkLimitOptions);
 
             Assert.NotNull(otlpLogRecord);
             Assert.Equal("Hello from tomato 2.99.", otlpLogRecord.Body.StringValue);
@@ -280,7 +284,7 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
             Assert.Single(logRecords);
 
             logRecord = logRecords[0];
-            otlpLogRecord = logRecord.ToOtlpLog();
+            otlpLogRecord = logRecord.ToOtlpLog(DefaultSdkLimitOptions);
             Assert.NotNull(otlpLogRecord);
             Assert.Equal("Hello from tomato 2.99.", otlpLogRecord.Body.StringValue);
 
@@ -305,7 +309,7 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
             var logger = loggerFactory.CreateLogger("OtlpLogExporterTests");
             logger.LogInformation("Log when there is no activity.");
             var logRecord = logRecords[0];
-            var otlpLogRecord = logRecord.ToOtlpLog();
+            var otlpLogRecord = logRecord.ToOtlpLog(DefaultSdkLimitOptions);
 
             Assert.Null(Activity.Current);
             Assert.True(otlpLogRecord.TraceId.IsEmpty);
@@ -335,7 +339,7 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
             }
 
             var logRecord = logRecords[0];
-            var otlpLogRecord = logRecord.ToOtlpLog();
+            var otlpLogRecord = logRecord.ToOtlpLog(DefaultSdkLimitOptions);
 
             Assert.Equal(expectedTraceId.ToString(), ActivityTraceId.CreateFromBytes(otlpLogRecord.TraceId.ToByteArray()).ToString());
             Assert.Equal(expectedSpanId.ToString(), ActivitySpanId.CreateFromBytes(otlpLogRecord.SpanId.ToByteArray()).ToString());
@@ -364,7 +368,7 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
             Assert.Single(logRecords);
 
             var logRecord = logRecords[0];
-            var otlpLogRecord = logRecord.ToOtlpLog();
+            var otlpLogRecord = logRecord.ToOtlpLog(DefaultSdkLimitOptions);
 
             Assert.NotNull(otlpLogRecord);
             Assert.Equal(logLevel.ToString(), otlpLogRecord.SeverityText);
@@ -415,7 +419,7 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
             Assert.Single(logRecords);
 
             var logRecord = logRecords[0];
-            var otlpLogRecord = logRecord.ToOtlpLog();
+            var otlpLogRecord = logRecord.ToOtlpLog(DefaultSdkLimitOptions);
 
             Assert.NotNull(otlpLogRecord);
             if (includeFormattedMessage)
@@ -434,7 +438,7 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
             Assert.Single(logRecords);
 
             logRecord = logRecords[0];
-            otlpLogRecord = logRecord.ToOtlpLog();
+            otlpLogRecord = logRecord.ToOtlpLog(DefaultSdkLimitOptions);
 
             Assert.NotNull(otlpLogRecord);
 
@@ -450,7 +454,7 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
             Assert.Single(logRecords);
 
             logRecord = logRecords[0];
-            otlpLogRecord = logRecord.ToOtlpLog();
+            otlpLogRecord = logRecord.ToOtlpLog(DefaultSdkLimitOptions);
 
             Assert.NotNull(otlpLogRecord);
 
@@ -473,7 +477,7 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
 
             var logRecord = logRecords[0];
             var loggedException = logRecord.Exception;
-            var otlpLogRecord = logRecord.ToOtlpLog();
+            var otlpLogRecord = logRecord.ToOtlpLog(DefaultSdkLimitOptions);
 
             Assert.NotNull(otlpLogRecord);
             var otlpLogRecordAttributes = otlpLogRecord.Attributes.ToString();
@@ -533,6 +537,53 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
             Assert.Equal("value2", request.ResourceLogs[0].ScopeLogs[1].Scope.Attributes[0].Value.StringValue);
             Assert.Equal(SemanticConventions.AttributeLogEventDomain, request.ResourceLogs[0].ScopeLogs[1].Scope.Attributes[1].Key);
             Assert.Equal("testLogger2EventDomain", request.ResourceLogs[0].ScopeLogs[1].Scope.Attributes[1].Value.StringValue);
+        }
+
+        [Fact]
+        public void CheckToOtlpLogRecordRespectsAttributeLimits()
+        {
+            var sdkLimitOptions = new SdkLimitOptions
+            {
+                AttributeCountLimit = 3, // 3 => LogCategory, exception.type and exception.message
+                AttributeValueLengthLimit = 8,
+            };
+
+            var logRecords = new List<LogRecord>();
+            using var loggerFactory = LoggerFactory.Create(builder =>
+            {
+                builder.AddOpenTelemetry(options =>
+                {
+                    options.AddInMemoryExporter(logRecords);
+                });
+            });
+
+            var logger = loggerFactory.CreateLogger("OtlpLogExporterTests");
+            logger.LogInformation(new NotSupportedException("I'm the exception message."), "Exception Occurred");
+
+            var logRecord = logRecords[0];
+            var otlpLogRecord = logRecord.ToOtlpLog(sdkLimitOptions);
+
+            Assert.NotNull(otlpLogRecord);
+            Assert.Equal(1u, otlpLogRecord.DroppedAttributesCount);
+
+            var exceptionTypeAtt = TryGetAttribute(otlpLogRecord, SemanticConventions.AttributeExceptionType);
+            Assert.NotNull(exceptionTypeAtt);
+
+            // "NotSuppo" == first 8 chars from the exception typename "NotSupportedException"
+            Assert.Equal("NotSuppo", exceptionTypeAtt.Value.StringValue);
+            var exceptionMessageAtt = TryGetAttribute(otlpLogRecord, SemanticConventions.AttributeExceptionMessage);
+            Assert.NotNull(exceptionMessageAtt);
+
+            // "I'm the " == first 8 chars from the exception message
+            Assert.Equal("I'm the ", exceptionMessageAtt.Value.StringValue);
+
+            var exceptionStackTraceAtt = TryGetAttribute(otlpLogRecord, SemanticConventions.AttributeExceptionStacktrace);
+            Assert.Null(exceptionStackTraceAtt);
+        }
+
+        private static OtlpCommon.KeyValue TryGetAttribute(OtlpLogs.LogRecord record, string key)
+        {
+            return record.Attributes.FirstOrDefault(att => att.Key == key);
         }
     }
 }